requirements.txt
pandas~=2.3.1
numpy~=2.3.2
matplotlib~=3.10.5
seaborn~=0.13.2
pillow~=11.3.0
PyQt6~=6.9.1
pip~=23.2.1
pytz~=2025.2
setuptools~=80.9.0
pyinstaller~=6.15.0
openpyxl~=3.1.5
scipy~=1.16.1
python-dateutil~=2.9.0.post0
altgraph~=0.17.4
six~=1.17.0
et_xmlfile~=2.0.0
contourpy~=1.3.3
fonttools~=4.59.0
packaging~=25.0
pefile~=2023.2.7
pyparsing~=3.2.3
cycler~=0.12.1
kiwisolver~=1.4.8
pywin32-ctypes~=0.2.3
pyinstaller-hooks-contrib~=2025.8